Reorganized folder structure and converted to single page layout
