+++ app/client/components/Dashboard.coffee
@@ -7,20 +7,27 @@ MotorList = require('./MotorList')
 PeripheralList = require('./PeripheralList')
 FinalCompPeripheralList = require('./FinalCompPeripheralList')
 Peripheral = require('./Peripheral')
-Environment = require('../../utils/Environment')
+Environment = require('../utils/Environment')
+DebugGamepads = require('./DebugGamepads')
 if Environment.isBrowser
+  Editor = require('./Editor')
   RobotActions = require('../actions/RobotActions')
 
 module.exports = Dashboard = React.createClass
   displayName: 'Dashboard'
   render: ->
+    editor = 'Loading...'
+    if Editor?
+      editor = <Editor/>
     return (
       <Grid fluid>
         <Row>
           <Col sm={8}>
+            { editor }
           </Col>
           <Col sm={4}>
             <FinalCompPeripheralList/>
+            <DebugGamepads/>
           </Col>
         </Row>
       </Grid>
