--- app/client/components/Edit.coffee
@@ -1,16 +1,0 @@
-React = require('react')
-RouteHandler = require('react-router').RouteHandler
-Environment = require('../utils/Environment')
-if Environment.isBrowser
-  Editor = require('./Editor')
-
-module.exports = Edit = React.createClass
-  displayName: 'Edit'
-  render: ->
-    content = 'Loading...'
-    if Editor?
-      content = <Editor/>
-    <div className="container">
-      { content }
-    </div>
-
