+++ app/client/components/FinalCompPeripheralList.js
