+++ app/client/components/PeripheralList.js
