+++ app/client/components/peripherals/BooleanSensor.js
