+++ app/client/components/peripherals/GenericPeripheral.js
