+++ app/client/components/peripherals/ScalarSensor.js
