+++ app/client/routes.coffee
@@ -4,14 +4,10 @@ DefaultRoute = Router.DefaultRoute
 Route = Router.Route
 
 App = require('./components/App')
-Edit = require('./components/Edit')
-Debug = require('./components/Debug')
-Dashboard = require('./dashboard/components/Dashboard')
+Dashboard = require('./components/Dashboard')
 
 module.exports = routes =
   <Route handler={App} path='/'>
     <DefaultRoute handler={Dashboard}/>
-    <Route handler={Edit} name='edit'/>
-    <Route handler={Debug} name='debug'/>
     <Route handler={Dashboard} name='dashboard'/>
   </Route>
